fix: BigInteger Etsy IDs, thread-safe sessions, and URL parsing
- Migrate Etsy shop_id/listing_id from Integer to BigInteger to handle
  large numeric IDs (>2^31)
- Refactor async engine/session with loop detection for multi-thread
  workers; same for Redis connection in TokenManager
- Eager-load Listing relationship via selectinload in SEO audit queries
- Support international Etsy URLs (hk-en/shop/NAME) in shop tracking
- Cap shop sync at 500 listings to prevent excessive pagination
- Protect against null fields in Etsy API listing data (tags, images, etc.)
- Add listing relation to SEOAudit frontend type

+++ backend/app/models/listing.py
@@ -1,6 +1,6 @@
 from decimal import Decimal
 
-from sqlalchemy import String, Integer, Float, Numeric, Text, ForeignKey
+from sqlalchemy import String, Integer, BigInteger, Float, Numeric, Text, ForeignKey
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db.base import Base, UUIDMixin, TimestampMixin
@@ -9,8 +9,8 @@ from app.db.base import Base, UUIDMixin, TimestampMixin
 class Listing(Base, UUIDMixin, TimestampMixin):
     __tablename__ = "listings"
 
-    listing_id: Mapped[int] = mapped_column(Integer, unique=True, index=True, nullable=False)
-    shop_id: Mapped[int] = mapped_column(Integer, ForeignKey("shops.shop_id"), nullable=False, index=True)
+    listing_id: Mapped[int] = mapped_column(BigInteger, unique=True, index=True, nullable=False)
+    shop_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("shops.shop_id"), nullable=False, index=True)
     title: Mapped[str] = mapped_column(Text, nullable=False)
     description: Mapped[str] = mapped_column(Text, nullable=True)
     tags: Mapped[list] = mapped_column(JSONB, default=list, server_default="[]")
