+++ backend/app/models/shop.py
@@ -1,6 +1,6 @@
 from decimal import Decimal
 
-from sqlalchemy import String, Integer, Float, Numeric
+from sqlalchemy import String, Integer, BigInteger, Float, Numeric
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db.base import Base, UUIDMixin, TimestampMixin
@@ -9,7 +9,7 @@ from app.db.base import Base, UUIDMixin, TimestampMixin
 class Shop(Base, UUIDMixin, TimestampMixin):
     __tablename__ = "shops"
 
-    shop_id: Mapped[int] = mapped_column(Integer, unique=True, index=True, nullable=False)
+    shop_id: Mapped[int] = mapped_column(BigInteger, unique=True, index=True, nullable=False)
     name: Mapped[str] = mapped_column(String(500), nullable=False)
     url: Mapped[str] = mapped_column(String(1000), nullable=False)
     total_listings: Mapped[int] = mapped_column(Integer, default=0, server_default="0")
